test(pallet-storage-provider-benchmarks): Change module conditionals to match market, in case we later move tests here
