+++ pallets/storage-provider/benchmarks/src/lib.rs
@@ -5,11 +5,14 @@ extern crate alloc;
 #[cfg(feature = "runtime-benchmarks")]
 pub mod benchmarking;
 
-#[cfg(all(feature = "runtime-benchmarks", test))]
-pub(crate) mod mock;
+#[cfg(test)]
+mod mock;
 
 #[cfg(any(feature = "runtime-benchmarks", test))]
 mod pallet;
 
 #[cfg(any(feature = "runtime-benchmarks", test))]
 pub use pallet::*;
+
+#[cfg(test)]
+mod test;
